+++ src/main/java/org/Eleks/Gmail/po/HomePage.java
@@ -1,15 +1,12 @@
 package org.Eleks.Gmail.po;
-
-import io.qameta.allure.Step;
-import org.Eleks.Gmail.bo.EmailSendPageBO;
-import org.Eleks.Gmail.factories.UserFactory;
+import org.Eleks.Gmail.utils.PropertyUtils;
 
 
 public class HomePage extends BasePage {
 
 //todo винести статичні методи в BO
     public HomePage() {
-        setExpectedUrl(UserFactory.getProperties("expectedUrlHomePage"));
+        setExpectedUrl(PropertyUtils.getProperties("expectedUrlHomePage"));
     }
 
 }
