+++ src/test/java/org/eleks/gmail/vstasiv/BaseTest.java
@@ -7,6 +7,7 @@ import org.Eleks.Gmail.factories.UserFactory;
 import org.Eleks.Gmail.listeners.AnnotationTransformer;
 import org.Eleks.Gmail.listeners.Retry;
 import org.Eleks.Gmail.listeners.TestListener;
+import org.Eleks.Gmail.utils.PropertyUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.WebDriver;
@@ -24,7 +25,7 @@ public class BaseTest  {
         WebDriver webDriver = DriverFactory.setUpWebDriver(DriverFactory.Browsers.CHROME);
         LOGGER.info("Browser is opened",BaseTest.class);
         webDriver.manage().window().maximize();
-        DriverFactory.getWebDriver().get(UserFactory.getProperties("testURL2"));
+        DriverFactory.getWebDriver().get(PropertyUtils.getProperties("testURL2"));
     }
     @AfterMethod
     @Step("Close the browser")
